Move the FailureCollector to configdomain
