+++ src/cmd/config.go
@@ -10,7 +10,6 @@ import (
 	"github.com/git-town/git-town/v11/src/domain"
 	"github.com/git-town/git-town/v11/src/execute"
 	"github.com/git-town/git-town/v11/src/git"
-	"github.com/git-town/git-town/v11/src/gohacks"
 	"github.com/spf13/cobra"
 )
 
@@ -62,7 +61,7 @@ func executeConfig(verbose bool) error {
 }
 
 func determineConfigConfig(run *git.ProdRunner) (ConfigConfig, error) {
-	fc := gohacks.FailureCollector{}
+	fc := configdomain.FailureCollector{}
 	branchTypes := run.GitTown.BranchTypes()
 	deleteOrigin := fc.ShipDeleteRemoteBranch(run.GitTown.ShouldShipDeleteOriginBranch())
 	giteaToken := run.GitTown.GiteaToken()
