+++ src/config/configdomain/failure_collector_test.go
@@ -1,4 +1,4 @@
-package gohacks_test
+package configdomain_test
 
 import (
 	"errors"
@@ -6,7 +6,6 @@ import (
 
 	"github.com/git-town/git-town/v11/src/config/configdomain"
 	"github.com/git-town/git-town/v11/src/domain"
-	"github.com/git-town/git-town/v11/src/gohacks"
 	"github.com/shoenig/test/must"
 )
 
@@ -16,7 +15,7 @@ func TestCollector(t *testing.T) {
 	t.Run("Bool", func(t *testing.T) {
 		t.Run("returns the given bool value", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			must.True(t, fc.Bool(true, nil))
 			must.False(t, fc.Bool(false, nil))
 			err := errors.New("test error")
@@ -26,7 +25,7 @@ func TestCollector(t *testing.T) {
 
 		t.Run("captures the first error it receives", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.Bool(true, nil)
 			fc.Bool(false, nil)
 			must.Nil(t, fc.Err)
@@ -39,7 +38,7 @@ func TestCollector(t *testing.T) {
 	t.Run("BranchesSyncStatus", func(t *testing.T) {
 		t.Run("returns the given value", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			syncStatuses := domain.BranchInfos{
 				{
 					LocalName:  domain.NewLocalBranchName("branch1"),
@@ -65,7 +64,7 @@ func TestCollector(t *testing.T) {
 
 		t.Run("captures the first error it receives", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.Bool(true, nil)
 			fc.Bool(false, nil)
 			must.Nil(t, fc.Err)
@@ -78,7 +77,7 @@ func TestCollector(t *testing.T) {
 	t.Run("Check", func(t *testing.T) {
 		t.Parallel()
 		t.Run("captures the first error it receives", func(t *testing.T) {
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.Check(nil)
 			must.Nil(t, fc.Err)
 			fc.Check(errors.New("first"))
@@ -86,7 +85,7 @@ func TestCollector(t *testing.T) {
 			must.ErrorContains(t, fc.Err, "first")
 		})
 		t.Run("indicates whether it received an error", func(t *testing.T) {
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			must.False(t, fc.Check(nil))
 			must.True(t, fc.Check(errors.New("")))
 			must.True(t, fc.Check(nil))
@@ -96,7 +95,7 @@ func TestCollector(t *testing.T) {
 	t.Run("Fail", func(t *testing.T) {
 		t.Parallel()
 		t.Run("registers the given error", func(t *testing.T) {
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.Fail("failed %s", "reason")
 			must.ErrorContains(t, fc.Err, "failed reason")
 		})
@@ -106,13 +105,13 @@ func TestCollector(t *testing.T) {
 		t.Parallel()
 		t.Run("returns the given HostingService value", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			must.EqOp(t, configdomain.HostingGitHub, fc.Hosting(configdomain.HostingGitHub, nil))
 			must.EqOp(t, configdomain.HostingGitLab, fc.Hosting(configdomain.HostingGitLab, errors.New("")))
 		})
 		t.Run("captures the first error it receives", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.Hosting(configdomain.HostingNone, nil)
 			must.Nil(t, fc.Err)
 			fc.Hosting(configdomain.HostingGitHub, errors.New("first"))
@@ -125,13 +124,13 @@ func TestCollector(t *testing.T) {
 		t.Parallel()
 		t.Run("returns the given SyncPerennialStrategy value", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			must.EqOp(t, configdomain.SyncPerennialStrategyMerge, fc.SyncPerennialStrategy(configdomain.SyncPerennialStrategyMerge, nil))
 			must.EqOp(t, configdomain.SyncPerennialStrategyRebase, fc.SyncPerennialStrategy(configdomain.SyncPerennialStrategyRebase, errors.New("")))
 		})
 		t.Run("captures the first error it receives", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.SyncPerennialStrategy(configdomain.SyncPerennialStrategyMerge, nil)
 			must.Nil(t, fc.Err)
 			fc.SyncPerennialStrategy(configdomain.SyncPerennialStrategyMerge, errors.New("first"))
@@ -144,13 +143,13 @@ func TestCollector(t *testing.T) {
 		t.Parallel()
 		t.Run("returns the given string value", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			must.EqOp(t, "alpha", fc.String("alpha", nil))
 			must.EqOp(t, "beta", fc.String("beta", errors.New("")))
 		})
 		t.Run("captures the first error it receives", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.String("", nil)
 			must.Nil(t, fc.Err)
 			fc.String("", errors.New("first"))
@@ -163,13 +162,13 @@ func TestCollector(t *testing.T) {
 		t.Parallel()
 		t.Run("returns the given string slice", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			must.Eq(t, []string{"alpha"}, fc.Strings([]string{"alpha"}, nil))
 			must.Eq(t, []string{"beta"}, fc.Strings([]string{"beta"}, errors.New("")))
 		})
 		t.Run("captures the first error it receives", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.Strings([]string{}, nil)
 			must.Nil(t, fc.Err)
 			fc.Strings([]string{}, errors.New("first"))
@@ -182,13 +181,13 @@ func TestCollector(t *testing.T) {
 		t.Parallel()
 		t.Run("returns the given SyncFeatureStrategy value", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			must.EqOp(t, configdomain.SyncFeatureStrategyMerge, fc.SyncFeatureStrategy(configdomain.SyncFeatureStrategyMerge, nil))
 			must.EqOp(t, configdomain.SyncFeatureStrategyRebase, fc.SyncFeatureStrategy(configdomain.SyncFeatureStrategyRebase, errors.New("")))
 		})
 		t.Run("captures the first error it receives", func(t *testing.T) {
 			t.Parallel()
-			fc := gohacks.FailureCollector{}
+			fc := configdomain.FailureCollector{}
 			fc.SyncFeatureStrategy(configdomain.SyncFeatureStrategyMerge, nil)
 			must.Nil(t, fc.Err)
 			fc.SyncFeatureStrategy(configdomain.SyncFeatureStrategyMerge, errors.New("first"))
